Assembler UMINUS

+++ tests/simple.c
@@ -1,10 +1,10 @@
 int main()
 {
-    float a = 2.0;
-    float b = 5.0;
-    float c;
+    int a = 2;
+    int b = 5;
+    int c;
 
-    c = (-a + b) * 2.0;
+    c = (-a + b) * 2;
     print("Hello world!");
     print(c);
 }
